+++ utils/scifigextract.py
